apps/users/views.py
from django.contrib.auth import get_user_model
from django.contrib.auth.mixins import LoginRequiredMixin
from django.urls import reverse
from django.views.generic import DetailView, RedirectView, UpdateView

from rest_framework import viewsets, mixins
from .serializers import UserSerializer

User = get_user_model()


class UserDetailView(LoginRequiredMixin, DetailView):

    model = User
    slug_field = "id"
    slug_url_kwarg = "id"


user_detail_view = UserDetailView.as_view()


class UserUpdateView(LoginRequiredMixin, UpdateView):

    model = User
    fields = ["image", "first_name", "last_name", "birthdate", "citizenship", "gender", "positions", "certifications"]

    def get_success_url(self):
        return reverse("users:detail", kwargs={"id": self.request.user.id})

    def get_object(self):
        return User.objects.get(id=self.request.user.id)


user_update_view = UserUpdateView.as_view()


class UserRedirectView(LoginRequiredMixin, RedirectView):

    permanent = False

    def get_redirect_url(self):
        return reverse("users:update")


user_redirect_view = UserRedirectView.as_view()
